fix seller test
