+++ seller-info-service/src/main/java/com/marketplace/sellerinfoservice/SellerInfoServiceApplication.java
@@ -1,0 +1,12 @@
+package com.marketplace.sellerinfoservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SellerInfoServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SellerInfoServiceApplication.class, args);
+    }
+}
