+++ seller-info-service/src/main/java/com/marketplace/sellerinfoservice/dto/SellerRegistration.java
@@ -1,0 +1,11 @@
+package com.marketplace.sellerinfoservice.dto;
+
+import com.marketplace.sellerinfoservice.validator.annotation.SellerEmailConstraint;
+import com.marketplace.sellerinfoservice.validator.annotation.SellerPasswordConstraint;
+public record SellerRegistration(
+        @SellerEmailConstraint
+        String email,
+        @SellerPasswordConstraint
+        String password
+) {
+}
