+++ seller-info-service/src/main/java/com/marketplace/sellerinfoservice/model/BusinessModel.java
@@ -1,0 +1,6 @@
+package com.marketplace.sellerinfoservice.model;
+
+public enum BusinessModel {
+    IP,
+    OOO
+}
