--- seller-info-service/src/main/java/com/yandex/market/sellerinfoservice/SellerInfoServiceApplication.java
@@ -1,12 +1,0 @@
-package com.yandex.market.sellerinfoservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SellerInfoServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SellerInfoServiceApplication.class, args);
-    }
-}
