--- seller-info-service/src/main/java/com/yandex/market/sellerinfoservice/controller/response/ErrorResponse.java
@@ -1,19 +1,0 @@
-package com.yandex.market.sellerinfoservice.controller.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorResponse {
-
-    private String statusCode;
-    private String message;
-    private LocalDateTime timestamp;
-}
