--- seller-info-service/src/main/java/com/yandex/market/sellerinfoservice/dto/SellerRegistration.java
@@ -1,11 +1,0 @@
-package com.yandex.market.sellerinfoservice.dto;
-
-import com.yandex.market.sellerinfoservice.validator.annotation.SellerEmailConstraint;
-import com.yandex.market.sellerinfoservice.validator.annotation.SellerPasswordConstraint;
-public record SellerRegistration(
-        @SellerEmailConstraint
-        String email,
-        @SellerPasswordConstraint
-        String password
-) {
-}
