--- seller-info-service/src/main/java/com/yandex/market/sellerinfoservice/model/BusinessModel.java
@@ -1,6 +1,0 @@
-package com.yandex.market.sellerinfoservice.model;
-
-public enum BusinessModel {
-    IP,
-    OOO
-}
