--- seller-info-service/src/test/resources/application-test.yaml
@@ -1,27 +1,0 @@
-#SPRING CONFIGURATION
-spring:
-  app:
-    seller:
-      url: /public/api/v1/sellers/
-      json-path: src/test/resources/json/
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:test;DB_CLOSE_ON_EXIT=FALSE;NON_KEYWORDS=VALUE
-    username: sa
-    password:
-  jpa:
-    hibernate:
-      ddl-auto: create
-    defer-datasource-initialization: true
-    show-sql: true
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.H2Dialect
-        format_sql: true
-
-#LOGGING CONFIGURATION
-logging:
-  level:
-    org:
-      hibernate:
-        sql: info
