+++ seller-info-service/src/test/resources/application-testcontainers.yaml
@@ -1,0 +1,23 @@
+spring:
+  app:
+    seller:
+      url: /public/api/v1/sellers
+      json-path: src/test/resources/json/
+  jpa:
+    properties:
+      hibernate:
+        show_sql: true
+        use_sql_comments: true
+        format_sql: true
+    hibernate:
+      ddl-auto: create-drop
+  datasource:
+    url: jdbc:tc:postgresql:14.1:///testdb
+    driver-class-name: org.testcontainers.jdbc.ContainerDatabaseDriver
+  flyway:
+    enabled: false
+
+
+logging:
+  level:
+    org.springframework: DEBUG
