--- seller-info-service/src/test/resources/application-testcontainers.yml
@@ -1,17 +1,0 @@
-spring:
-  app:
-    seller:
-      url: /public/api/v1/sellers
-  jpa:
-    properties:
-      hibernate:
-        show_sql: true
-        use_sql_comments: true
-        format_sql: true
-    hibernate:
-      ddl-auto: create-drop
-  datasource:
-    url: jdbc:tc:postgresql:14.1:///testdb
-    driver-class-name: org.testcontainers.jdbc.ContainerDatabaseDriver
-  flyway:
-    enabled: false
