--- seller-info-service/src/test/resources/db/insert_tests_fields.sql
@@ -1,12 +1,0 @@
-INSERT INTO sellers (id, external_id, first_name, last_name, email, legal_address, company_name, image_url,
-                     business_model, ITN, PSRN, BIC, payment_account, corporate_account, is_deleted, password)
-VALUES (10001, '37678201-f3c8-4d5c-a628-2344eef50c51', 'Новое имя', 'Новая фамилия', 'newemail@mail.ru', 'Ул Новая, 1',
-        'Новая компания', 'new.png', 'IP', '123456789', '123456', '1234', 'ACCOUNT123456', 'SACCOUNT123', false, 'a123ddd!A'),
-       (10002, '37678201-f3c8-4d5c-a628-2344eef50c52', 'Имя 2', 'Фамилия 2', 'email2@mail.ru', 'Ул 2',
-        'Компания 2', 'url2.png', 'IP', '1111111111', '222222', '3333', 'ACCOUNT222222', 'SACCOUNT222', false,'a123ddd!A1'),
-       (10003, '37678201-f3c8-4d5c-a628-2344eef50c53', 'Имя 3', 'Фамилия 3', 'email3@mail.ru', 'Ул 3',
-        'Компания 3', 'url3.png', 'IP', '4444444444', '555555', '6666', 'ACCOUNT333333', 'SACCOUNT333', false, 'a123ddd!A2'),
-       (10004, '37678201-f3c8-4d5c-a628-2344eef50c54', 'Имя 4', 'Фамилия 4', 'email4@mail.ru', 'Ул 4',
-        'Компания 4', 'url4.png', 'IP', '7777777777', '888888', '9999', 'ACCOUNT444444', 'SACCOUNT444', false, 'a123ddd!A3'),
-       (10005, '37678201-f3c8-4d5c-a628-2344eef50c55', 'Имя 5', 'Фамилия 5', 'email5@mail.ru', 'Ул 5',
-        'Компания 5', 'url5.png', 'IP', '0000000000', '111111', '2222', 'ACCOUNT555555', 'SACCOUNT555', false, 'a123ddd!A4');
