--- seller-info-service/src/test/resources/db/truncate.sql
@@ -1,1 +1,0 @@
-truncate table sellers;
