+++ seller-info-service/src/test/resources/json/create/create_seller_request_dto.json
@@ -1,4 +1,4 @@
 {
-  "password": "Zhi123!asd",
-  "email": "alex@mail.ru"
+  "password": "Aa4321!8ee",
+  "email": "arkady@gmail.ru"
 }
